add contract order query

+++ src/main/java/com/fota/trade/domain/ContractOrderDO.java
@@ -21,8 +21,8 @@ public class ContractOrderDO {
     private Integer operateType;
     private Integer operateDirection;
     private Integer lever;
-    private BigDecimal totalAmount;
-    private BigDecimal unfilledAmount;
+    private Long totalAmount;
+    private Long unfilledAmount;
     private Integer closeType;
     private BigDecimal price;
     private BigDecimal fee;
